Add restart: always to docker-compose example
Twint will sometimes return an error and crash the container stating that it cannot find a token.  Simplest fix is to just get docker to restart the container as there's nothing inherently wrong other than to try again.
A code fix may be considered in the future to handle the error case.